% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitreg.R
\name{check_and_prepare_newdata}
\alias{check_and_prepare_newdata}
\title{Check and Prepare Newdata}
\usage{
check_and_prepare_newdata(newdata, object)
}
\arguments{
\item{newdata}{data.frame with the \code{newdata} for the prediction.}

\item{mf}{the model frame of the model (i.e., training data).}
}
\value{
Returns a modified version of \code{newdata} (a subset) or throws
errors if covariates are missing, or a class/level mismatch is found.
}
\description{
If \code{newdata} are provided when calling the predict method, this function
is used to (i) reduce \code{newdata} to the covariates originally used to
estimate the model, and if factor checks all levels are known (i.e.,
all levels have been seen in the training data).
TODO(R): This ensures we do not forward missing values to C which (currently)
would result in a segfault as NA handling is not properly implemented -- though
it already checks for it and throws errors).
}
\author{
Reto
}
