% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{get_elementwise_colnames}
\alias{get_elementwise_colnames}
\title{Create Column Names for Return Matrix}
\usage{
get_elementwise_colnames(
  x,
  prefix = NULL,
  digits = pmax(3L, getOption("digits") - 3L)
)
}
\arguments{
\item{x}{numeric, thresholds (pdf, cdf) or probabilities (quantile).}

\item{prefix}{If \code{NULL} (quantiles) the result is in percent,
else this prefix is used for each 'x'.}

\item{digits}{Integer, number of significant digits for names.}
}
\description{
Used in the S3 methods pdf, cdf, and quantile. If \code{elementwise = FALSE}
the return is a matrix; this helper function creates the names based
on the thresholds/probabilities used.
}
