% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitreg.R
\name{make_breaks}
\alias{make_breaks}
\title{Create Breaks for (Pseudo-)bins}
\usage{
make_breaks(y, breaks = 30, scale = FALSE, ...)
}
\arguments{
\item{y}{numeric vector with response data.}

\item{breaks}{number of breaks to be created (single integer).}

\item{scale}{logical, scaling involved?}
}
\value{
Returns a numeric vector with the breaks.
}
\description{
Calculates the breaks (point intersection between breaks)
which span the range of \code{y}.
}
