% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{num2bin}
\alias{num2bin}
\title{Convert Numerics to Pseudo-Bins}
\usage{
num2bin(
  x,
  breaks = NULL,
  censored = c("uncensored", "left", "right", "both"),
  verbose = FALSE
)
}
\arguments{
\item{x}{numeric vector.}

\item{breaks}{numeric vector with point intersections to create the bins.}

\item{censored}{character, defines if the (pseudo-)bins are censored or not
(see 'Details' section).}

\item{verbose}{logical, if \code{TRUE} some messages are printed if needed
(defaults to \code{FALSE}).}
}
\description{
Convert response on original scale to (pseudo-)bin index.
}
\details{
Converts the numeric values in \code{x} into (pseudo-)bins by cutting the data
into \code{length(breaks) - 1} segments. If \code{censored = "uncensored"} (no censoring),
data outside \code{range(break)} will be set to \code{NA_integer_}.
if \code{x < min(breaks)} or \code{length(breaks)} if \code{x > max(breaks)}.

If \code{censored = "left"} all values \code{x <= min(breaks)} are assigned to as bin \code{0L},
followed by bin \code{1L} for \verb{(breaks[1], breaks[2])}, \code{2L} for \verb{[breaks[2], breaks[3])} etc.
Thus, the highest bin containing data is \code{length(breaks)} as well.

If \code{censored = "right"} all values \code{x >= max(breaks)} are set to the highest bin
\code{length(breaks)}, those \verb{< max(breaks)} are still set \code{NA_integer_}.

When \code{censored = "both"} the rules above are combined. Due to the necessity that
we need tone dedicated 'censored bin index' on the left (\code{0L}), the right most
upper bin is shifted by \code{+1} compared to when we only use \code{censored = "right"}.

Less important to know for an end-user as this is mainly handelled internally.
}
\examples{
x      <- 0:10
breaks <- c(2, 6, 8)
transitreg:::num2bin(x, breaks = breaks)
transitreg:::num2bin(x, breaks = breaks, censored = "left")
transitreg:::num2bin(x, breaks = breaks, censored = "right")
transitreg:::num2bin(x, breaks = breaks, censored = "both")
transitreg:::num2bin(x, breaks = breaks, censored = "both")

}
