% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Shannon}
\alias{Shannon}
\title{Weather Data Shannon Airport (Ireland)}
\format{
A data frame containing historical meteorological observations from
Shannon Airport (Clare County, Ireland) spanning over 79 years, without gaps
or missing values. Observations are valid from 0000 UTC to 0000 UTC.

The data frame contains the following variables:
\itemize{
\item \code{date}: Date of observation
\item \code{maxtp}: Daily maximum temperature [°C]
\item \code{mintp}: Daily minimum temperature [°C]
\item \code{rain}: Daily precipitation sum [mm]
\item \code{sun}: Sunshine duration [hours]
}

For more details about the observations and recording period,
visit \url{https://www.met.ie/climate/what-we-measure}.
}
\source{
Met Éireann, the Irish Meteorological Service,
licensed under the Creative Commons Attribution 4.0 International
(CC BY 4.0) license via. Available at
\url{https://www.met.ie/climate/available-data/historical-data}.
}
\usage{
Shannon
}
\description{
This data set contains daily meteorological observations from Shannon Airport,
located on the west coast of Ireland, covering the period from September 1,
1945, to January 31, 2025 (29,008 days, approximately 79.5 years). The data,
provided by the Irish Meteorological Service (Met Éireann), includes daily
minimum and maximum temperatures, precipitation amounts, and sunshine duration.
}
\details{
\strong{Station information}
\itemize{
\item Country/county: Ireland, Clare
\item Station: Shannon Airport (ID 518)
\item Altitude: 15 m a.m.s.l.
\item Location: 52.69028°N, 8.91806°W
}

For more details, see the \href{https://wow.met.ie/}{weather observations website}.
}
\examples{
data(Shannon)
summary(Shannon)

}
\author{
Reto
}
\keyword{datasets}
