% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{survival}
\alias{survival}
\alias{as.character.survival}
\alias{print.survival}
\alias{format.survival}
\alias{[.survival}
\title{Creating Survival Object}
\usage{
survival(time, event)

\method{as.character}{survival}(x, ...)

\method{print}{survival}(x, quote = FALSE, ...)

\method{format}{survival}(x, ...)

\method{[}{survival}(x, i, ...)
}
\arguments{
\item{time}{the time to event, the varaible the binning is taking place on.}

\item{event}{must evaluate to logical, whether or not an event happened
(\code{FALSE} equals survived, \code{TRUE} equals dead). If \code{event} is a single
logical it is recycled for all observations, else the length of
the two vectors \code{time} and \code{event} must match exactely.}
}
\value{
An object of class \code{survival} which can be used as response
in transitreg models.
}
\description{
Creates a survival object used as response variable in a transitreg
model formula.
}
\author{
Reto
}
