% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response_and_data.R
\name{transitreg_response}
\alias{transitreg_response}
\title{Building Model Response for Transition Model}
\usage{
transitreg_response(x, response, breaks, verbose = FALSE, ...)
}
\arguments{
\item{x}{model framme, a \code{data.frame} where the first variable is expected
to be the model response.}

\item{response}{Character of length one, name of the response variable.}

\item{breaks}{numeric vector with breaks for discretizing the original response.
Can be \code{NULL} if no discretization is required.}

\item{verbose}{Logical value indicating whether information about the transformation
process should be printed to the console. Default is \code{FALSE}.}

\item{\dots}{Additional options forwarded from \verb{[transitreg()]}.}
}
\value{
Named list with the following three elements:
\itemize{
\item \code{mf}: Modified version of input \code{x}; altered response with (pseudo-)bins
if required.
\item \code{breaks}: Numeric vector with breaks, point intersection between (pseudo-)bins.
\item \code{bins}: Number of bins, same as \code{length(breaks) - 1L}.
\item \code{ym}: mid-point of the bins.
\item \code{yc}: index vector, (pseudo-)bin index/number.
}
}
\description{
To estimate a model as well as to perform predictions the model frame needs
to be built. This function (for internal use only) sets up the model
response properly, scaling and discretizising the response (if requested).
}
