% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{make_breaks}
\alias{make_breaks}
\title{Create Breaks for (Pseudo-)bins}
\usage{
make_breaks(y, breaks, censored)
}
\arguments{
\item{y}{a numeric vector with response data.}

\item{breaks}{a single numeric (integer) >= 1, number of breaks to be created.}

\item{censored}{character, one of \code{"uncensored"}, \code{"left"}, \code{"right"}.}
}
\value{
Returns a numeric vector with the breaks.
}
\description{
Calculates the breaks (point intersection between breaks)
which span the range of \code{y} if the user did not specify
a vector of breaks, but only a single numeric (integer).
}
