% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitreg.R
\name{plot.transitreg}
\alias{plot.transitreg}
\title{Plot Method for Transition Model Fits}
\usage{
\method{plot}{transitreg}(x, which = "effects", ask = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{"transitreg"} resulting from a call to \code{\link[=transitreg]{transitreg()}}}

\item{which}{A character or integer, specifying the type of plot(s) to
generate (See 'Details').}

\item{ask}{Either \code{NULL}, \code{TRUE} or \code{FALSE}. If \code{NULL} it will evaluate to
\code{TRUE} if more than one plot is requested via \code{which}.}

\item{\dots}{Additional arguments passed to the underlying plotting functions.}
}
\value{
Returns \code{NULL} invisibly. Generates plots as a side effect.
}
\description{
Provides diagnostic and visualization plots for transition models fitted using
the \code{\link[=transitreg]{transitreg()}} function. The method supports plotting effects and quantile
residual diagnostic plots.
}
\details{
The function allows to control what to plot via the argument \code{which}.
Options include:
\itemize{
\item \code{"effects"} Effects of the predictors on the response.
Requires that the model is estimated by \code{\link[mgcv:gam]{mgcv::gam()}} or \code{\link[mgcv:bam]{mgcv::bam()}}.
\item \code{"rootogram"}: Rootogram for assessing goodness of fit.
\item \code{"qqrplot"}: Q-Q plot for quantile residuals.
\item \code{"wormplot"}: Worm plot for quantile residuals.
\item \code{"pithist"}: Probability integral transform (PIT) histogram.
}

Multiple options can be specified as a character vector or numeric indices.

The \code{\link[=plot.transitreg]{plot.transitreg()}} method provides flexible visualization options for
evaluating transition model fits. Users can choose to:
\itemize{
\item Visualize the effects of predictors on the response variable
(if the model is a GAM, see \code{\link[mgcv:gam]{mgcv::gam()}}).
\item Evaluate quantile residuals through histograms, Q-Q plots, or worm plots
employing the \code{topmodels} framework for model evaluation.
}

The \code{which} argument controls the type of plots generated. By default, the
\code{"effects"} plot is shown if the model supports it. Residual-based plots
(\code{"rootogram"}, \code{"qqrplot"}, \code{"wormplot"}, \code{"pithist"}) provide insights into
the goodness of fit of the model and model calibration.
}
\examples{
## Example: Fit a transition model and generate plots.
set.seed(123)
n <- 500
x <- runif(n, -3, 3)
y <- rpois(n, exp(2 + sin(x)))
b <- transitreg(y ~ s(theta) + s(x))

## Plot effects.
plot(b, which = "effects")

## Plot residuals.
plot(b, which = c("rootogram", "pithist", "qqrplot", "wormplot"))

## Custom plot layout.
par(mfrow = c(2, 2))
plot(b, which = 2:5, ask = FALSE)

}
\seealso{
\code{\link[=transitreg]{transitreg()}}, \code{\link[=predict.transitreg]{predict.transitreg()}}.
}
\keyword{methods}
\keyword{models}
\keyword{visualization}
