% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rainIreland}
\alias{rainIreland}
\title{Rain Ireland}
\format{
A data frame containing xyz
\itemize{
\item \code{date}: The observations are valid from
0900 UTC of the previous day until 0900 UTC of the date provided.
\item \code{rain}: 24 hour precipitation sum in mm.
}

For details regarding the observations and recording period please
see 'Rainfall Stations' on \url{https://www.met.ie/climate/what-we-measure}.
}
\source{
Met Éireann, the Irish Meteorological Service,
licensed under the Creative Commons Attribution 4.0 International
(CC BY 4.0) license via. Available at
\url{https://www.met.ie/climate/available-data/historical-data}.
}
\usage{
rainIreland
}
\description{
See also \href{https://wow.met.ie/}{weather observations website}.
}
\details{
Station information:
\itemize{
\item Country/county: Ireland, Limerick
\item Station: Patrickswell (Dooneen; ID 4811)
\item Altitude (m): \code{27}
\item Location (deg): \code{52.59500}N/\code{-8.67278}E
}
}
\examples{
data(rainIreland)
hist(sqrt(rainIreland$rain), freq = FALSE,
     main = "Daily precipitation sums\nPatrickswell, Ireland",
     xlab = expression(sqrt(rain)), col = "steelblue")

}
\author{
Reto
}
\keyword{datasets}
