% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Shannon}
\alias{Shannon}
\title{Weather Shannon Airport (Ireland)}
\format{
Data frame with historical meteorological observations
at Shannon airport (Clare county, Ireland). Observations are
valid between 0000 UTC and 0000 UTC.
\itemize{
\item \code{date}: Date of observation.
\item \code{maxtp}: Daily maximum temperature \link{degrees Celsius}.
\item \code{mintp}: Daily minimum temperature \link{degrees Celsius}.
\item \code{rain}: Daily precipitation sum \link{millimeters}.
\item \code{sun}: Sunshine duration \link{hours}
}

For details regarding the observations and recording period please
see \url{https://www.met.ie/climate/what-we-measure}.
}
\source{
Met Éireann, the Irish Meteorological Service,
licensed under the Creative Commons Attribution 4.0 International
(CC BY 4.0) license via. Available at
\url{https://www.met.ie/climate/available-data/historical-data}.
}
\usage{
Shannon
}
\description{
See also \href{https://wow.met.ie/}{weather observations website}.
}
\details{
Station information:
\itemize{
\item Country/county: Ireland, Clare
\item Station: Shannon Airport (ID 518)
\item Altitude (m): \code{15}
\item Location (deg): \code{52.69028}N/\code{-8.91806}E
}
}
\examples{
data(Shannon)
summary(Shannon)

}
\author{
Reto
}
\keyword{datasets}
