% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transition.R
\name{Transition}
\alias{Transition}
\alias{dtransit}
\alias{ptransit}
\alias{qtransit}
\alias{rtransit}
\alias{c.Transition}
\alias{as.matrix.Transition}
\alias{pdf.Transition}
\alias{log_pdf.Transition}
\alias{cdf.Transition}
\alias{quantile.Transition}
\alias{median.Transition}
\alias{mean.Transition}
\alias{random.Transition}
\alias{is_discrete.Transition}
\alias{is_continuous.Transition}
\alias{support.Transition}
\alias{plot.Transition}
\title{Creates a Transition Distribution}
\usage{
Transition(x, breaks, discrete = NULL)

dtransit(x, d, log = FALSE, ncores = NULL)

ptransit(x, d, lower.tail = TRUE, log.p = FALSE, ncores = NULL)

qtransit(p, d, lower.tail = TRUE, log.p = FALSE, ncores = NULL)

rtransit(n, d, ncores = NULL)

\method{c}{Transition}(...)

\method{as.matrix}{Transition}(x, expand = FALSE, ...)

\method{pdf}{Transition}(d, x, drop = TRUE, elementwise = NULL, ncores = NULL, ...)

\method{log_pdf}{Transition}(d, x, drop = TRUE, elementwise = NULL, ncores = NULL, ...)

\method{cdf}{Transition}(d, x, drop = TRUE, elementwise = NULL, ncores = NULL, ...)

\method{quantile}{Transition}(x, probs, drop = TRUE, elementwise = NULL, ncores = NULL, ...)

\method{median}{Transition}(x, na.rm = NULL, ncores = NULL, ...)

\method{mean}{Transition}(x, ncores = NULL, ...)

\method{random}{Transition}(x, n = 1L, drop = TRUE, ...)

\method{is_discrete}{Transition}(d, ...)

\method{is_continuous}{Transition}(d, ...)

\method{support}{Transition}(d, drop = NULL, ...)

\method{plot}{Transition}(x, type = c("tp", "cdf", "pdf"), all = FALSE, n = 8L, ...)
}
\arguments{
\item{x}{Numeric, vector of quantiles.}

\item{breaks}{numeric vector of points of intersection of the breaks.
The length the vector must be of \code{length(x) + 1} (if \code{x} is
a vector) or \code{ncol(x) + 1} if \code{x} is a matrix. Must be
monotonically increasing.}

\item{discrete}{\code{NULL} or logical. If \code{NULL} and the breaks look like
count data breaks (\code{c(-0.5, 0.5, 1.5, ...)}) discrete will be set
\code{TRUE}, else \code{FALSE}.}

\item{d}{An object of class \code{Transition}.}

\item{log, log.p}{Logical, if \code{TRUE}, probabilities \code{p} are given as \code{log(p)}.}

\item{ncores}{Number of cores to be used (see \code{\link[=transitreg]{transitreg()}} for details).}

\item{lower.tail}{Logical, if \code{TRUE} (default), probabilities are
\code{P[X <= x]} otherwise, \code{P[X > x]}.}

\item{p}{Numeric, vector of probabilities.}

\item{n}{Integer. Maximum number of distributions to be plotted, defaults to 8.}

\item{\dots}{unused.}

\item{expand}{logical, if FALSE (default) the wide format is
returned, else the extended (long) form.}

\item{drop}{Logical. Should the result be simplified to a vector if possible?}

\item{elementwise}{Logical. Should each distribution in \code{x} be evaluated at
all elements of \code{probs} (\code{elementwise = FALSE}, yielding a matrix)? Or, if
\code{x} and \code{probs} have the same length, should the evaluation be done element
by element (\code{elementwise = TRUE} yielding a vector)? The default of \code{NULL}
means that \code{elementwise = TRUE} is used if the lengths match and otherwise
\code{elementwise = FALSE} is used.}

\item{probs}{numeric vector of probabilities with values in \verb{[0,1]}.
(Values up to ‘2e-14’ outside that range are accepted and
moved to the nearby endpoint.) TODO(R): SURE?}

\item{na.rm}{Unused.}

\item{type}{Character, type of plot to be created. Either
\code{"tp"} (transition probabilities), \code{"cdf"} or \code{"pdf"}.}

\item{all}{Logical. If \code{TRUE} all distributions in \code{x} will be drawn.
Else only the first N ones (default is 8).}
}
\value{
Returns an object of class \code{c("Transition", "distribution")}.
TODO(R): Missing.

Numeric matrix. If \code{expand = FALSE} the return is of dimension
\code{c(length(x), <ncol>)}.

If \code{expand = TRUE} the returned matrix is of dimension
\code{c(length(x) * <ncol>, 3L)} where the four columns contain
\code{index} (1, ..., length(x)) where each index corresponds to the
row-index of the original input \code{x} (distribution identifier),
\code{theta} which is the 'bin' the transition probability belongs to,
followed by the transition probabilities \code{tp} itself.
This expanded version is used when calling the .C functions.
}
\description{
A 'Transition' distrubiton consists of a series of \code{K} transition
probabilities (TP) for \code{K} 'bins' (counts or pseudo-counts).
}
\details{
TODO(R): Missing
}
\examples{
# TODO(R) Write example

}
\author{
Reto
}
\concept{Transition distribution}
\concept{distributions}
