% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transitreg.R
\name{transitreg_detect_cores}
\alias{transitreg_detect_cores}
\title{Detect number of cores for OpenMP}
\usage{
transitreg_detect_cores(verbose = TRUE)
}
\arguments{
\item{verbose}{logical, if \code{TRUE} a message is shown.}
}
\value{
Number of available cores (integer). If OpenMP is not
available, \code{1L} is returned.
}
\description{
The calculation of CDFs and PDFs is implemented in C and allows
for parallelization using OpenMP. This function detects how may
cores are available in total (if OpenMP is available).
}
\author{
Reto
}
