% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R
\name{transitreg_glmnet}
\alias{transitreg_glmnet}
\alias{predict.transitreg_glmnet}
\title{Estimat Transition Probabilities via \code{glmnet}}
\usage{
transitreg_glmnet(formula, data, nfolds = 10, ...)

\method{predict}{transitreg_glmnet}(object, type = "response", s = "lambda.min", ...)
}
\arguments{
\item{formula}{An object of class \code{formula}.}

\item{data}{A data.frame containing the required data to fit a binomial
regression model (given \code{formula}) using \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}}.}

\item{nfolds}{Integer, defaults to \code{10L}. Number of cross-folds for the
glmnet model.}

\item{\dots}{currently unused.}

\item{object}{an object of class \code{transitreg_glmnet}.}

\item{type}{character, defaults to \code{"response"}.}

\item{s}{value(s) of the penalty parameter, defaults to \code{"lambda.min"}.
See \code{\link[glmnet:predict.glmnet]{glmnet::predict.glmnet()}} for details.}
}
\description{
Experimental engine used to estimate the transition probabilities
in a \code{\link[=transitreg]{transitreg()}} model.
}
